# src/main/java/com/janfer/estoque/domain/dtos/ProdutoEntradaPostDTO.java
package com.janfer.estoque.domain.dtos;

import com.fasterxml.jackson.annotation.JsonFormat;
import com.janfer.estoque.domain.entities.ProdutoCapa;
import lombok.Data;

import java.io.Serial;
import java.io.Serializable;
import java.time.Instant;
import java.util.Date;

@Data
public class ProdutoEntradaPostDTO implements Serializable {

  @Serial
  private static final long serialVersionUID = 1L;
  private Long numeroNota;
  @JsonFormat(pattern = "dd/MM/yyyy")
  private Date dataPedido;
  @JsonFormat(pattern = "dd/MM/yyyy")
  private Date dataEntrega;
  private Double precoCompra;
  private Long quantidade;
  private String observacao;
  private transient ProdutoCapa produtoCapa;

}